+++ src/main/java/com/drmangotea/tfmg/config/TFMGServerConfig.java
@@ -6,6 +6,7 @@ import com.drmangotea.tfmg.config.server.TFMGStress;
 import net.createmod.catnip.config.ConfigBase;
 
 public class TFMGServerConfig extends ConfigBase {
+
 	public final TFMGStress stressValues = nested(0, TFMGStress::new, "Fine tune the kinetic stats of individual components");
 	public final TFMGResistivity resistivityValues = nested(0, TFMGResistivity::new, "Fine tune the resistivity stats of individual cable types");